backend/python/src/auth/requirements.txt
astroid==3.3.8
blinker==1.9.0
click==8.1.8
dill==0.3.9
Flask==3.1.0
Flask-MySQLdb==2.0.0
importlib_metadata==8.6.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.7
parso==0.8.4
platformdirs==4.3.6
PyJWT==2.10.1
pylint==3.3.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
